feat: add gradient, visual
