+++ assets/css/styles.css
@@ -1,8 +1,75 @@
+@import url('https://fonts.googleapis.com/css?family=Dosis:400,700&display=swap');
+body {
+    background: linear-gradient(232deg, #3391a6, #25594a, #3fa663, #2d7345, #262626);
+    background-size: 1000% 1000%;
+    animation: bg-grad 32s ease infinite;
+}
+
+@keyframes bg-grad {
+    0%{background-position:96% 0%}
+    50%{background-position:5% 100%}
+    100%{background-position:96% 0%}
+}
+
 .main {
     width: 1024px;
     margin: 0 auto;
+    font-family: 'Dosis', sans-serif;
+}
+
+/* HEADER */
+
+img {
+    display: block;
+    margin: 0 auto;
+}
+
+/* MAIN */
+
+.form-output {
+    display: grid;
+    grid-template-columns: 200px 300px 300px 200px;
+    grid-template-rows: 50px 100px 200px;
+    grid-gap: 20px;
+}
+
+.form-output select {
+    grid-column: 2;
+    place-self: center;
+}
+
+.form-output .cat-image {
+    grid-column: 3;
+    place-self: center;
+}
+
+.form-output .out {
+    grid-row: 3;
+    grid-column: 2 / 4;
+}
+
+textarea {
+    resize: vertical;
+    overflow: auto;
+    grid-row: 2;
+    grid-column: 2 / 4;
+}
+
+h1 {
+    visibility: hidden;
 }
 
 .output {
     overflow: auto;
+}
+
+/* FOOTER */
+
+footer {
+    position: absolute;
+    bottom: 0;
+    width: 1024px;
+    height: 50px;
+    font-size: 20px;
+    text-align: center;
 }
