--- internalerror.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Internal Server Error</title>
-</head>
-<body>
-    <h1>500 Internal Server Error</h1>
-    <p>Oops! Something Wrong</p>
-    <a href="http://localhost:8080/">Back to Magic Cat</a>
-</body>
-</html>
