+++ main.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"net/http"
 
-	"./ascii"
+	ascii "./app"
 )
 
 type ExecOutput struct {
@@ -24,7 +24,7 @@ func ValidAscii(s string) bool {
 
 func internalServerError(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(http.StatusInternalServerError)
-	t, _ := template.ParseFiles("internalerror.html")
+	t, _ := template.ParseFiles("error/internalerror.html")
 	err := t.Execute(w, nil)
 	if err != nil {
 		log.Fatal(err)
@@ -44,7 +44,7 @@ func Handler(w http.ResponseWriter, r *http.Request) {
 			r.ParseForm()
 			if !ValidAscii(r.Form.Get("input")) {
 				w.WriteHeader(http.StatusBadRequest)
-				t, err := template.ParseFiles("badrequest.html")
+				t, err := template.ParseFiles("error/badrequest.html")
 				if err != nil {
 					internalServerError(w, r)
 				}
@@ -52,7 +52,7 @@ func Handler(w http.ResponseWriter, r *http.Request) {
 			} else {
 				output, status := ascii.AsciiOutput(r.Form["input"][0], r.Form["font"][0])
 				log.Printf("method: %v / font: %v / input: %v / statuscode: %v\n", r.Method, r.Form["font"][0], r.Form["input"][0], status)
-				if status > 200 {
+				if status == 500 {
 					internalServerError(w, r)
 				} else {
 					ex := ExecOutput{
@@ -70,7 +70,7 @@ func Handler(w http.ResponseWriter, r *http.Request) {
 		}
 	} else {
 		w.WriteHeader(http.StatusNotFound)
-		t, err := template.ParseFiles("notfound.html")
+		t, err := template.ParseFiles("error/notfound.html")
 		if err != nil {
 			internalServerError(w, r)
 			return
